Style app to match design
No CSS framework has been used, as the app is very simple and the
cost of overhead of maintaining the dependency and the increased page
load size does not seem worth the negligible benefit.

No web fonts have been added, to keep the overhead of page load to a
minimum. Close substitutes have been chosen, although they need testing
and tweaking on Android and Windows.